+++ R/FullConditionals.R
@@ -16,20 +16,20 @@
       u.eta      <- diag(Q) + crossprod(load.psi, lmat)
       u.eta      <- if(Q1) sqrt(u.eta) else .chol(u.eta)
       mu.eta     <- c.data %*% (load.psi %*% if(Q1) 1/(u.eta * u.eta) else chol2inv(u.eta))
-        mu.eta    + t(backsolve(u.eta, matrnorm(Q, N)))
+        mu.eta    + t(backsolve(u.eta, matrnorm(Q, N), k=Q))
     }
 
   # Loadings
     .sim_load    <- function(l.sigma, Q, c.data, eta, psi.inv, EtE, Q1)  {
       u.load     <- l.sigma  + psi.inv * EtE
       u.load     <- if(Q1) sqrt(u.load) else .chol(u.load)
-        psi.inv   * (if(Q1) 1/(u.load  * u.load) else chol2inv(u.load)) %*% crossprod(eta, c.data) + backsolve(u.load, stats::rnorm(Q))
+        psi.inv   * (if(Q1) 1/(u.load  * u.load) else chol2inv(u.load)) %*% crossprod(eta, c.data) + backsolve(u.load, stats::rnorm(Q), k=Q)
     }
 
     .sim_load_s  <- function(Q, c.data, eta, phi, tau, psi.inv, EtE, Q1, sigma = 1L) {
       u.load     <- diag(phi * tau * sigma, Q) + psi.inv * EtE
       u.load     <- if(Q1) sqrt(u.load) else .chol(u.load)
-        psi.inv   * (if(Q1) 1/(u.load  * u.load) else chol2inv(u.load)) %*% crossprod(eta, c.data) + backsolve(u.load, stats::rnorm(Q))
+        psi.inv   * (if(Q1) 1/(u.load  * u.load) else chol2inv(u.load)) %*% crossprod(eta, c.data) + backsolve(u.load, stats::rnorm(Q), k=Q)
     }
 
   # Uniquenesses
@@ -63,22 +63,22 @@
     }
 
   # Local Shrinkage
-    .sim_phi     <- function(Q, P, nu1, nu2, tau, load.2, sigma = 1L) {
-        matrix(stats::rgamma(P * Q, shape=nu1 + 0.5, rate=nu2 + (sigma * sweep(load.2, 2L, tau, FUN="*", check.margin=FALSE))/2), nrow=P, ncol=Q)
+    .sim_phi     <- function(Q, P, nu1.5, nu2, tau, load.2, sigma = 1L) {
+        matrix(stats::rgamma(P * Q, shape=nu1.5, rate=nu2 + (sigma * sweep(load.2, 2L, tau, FUN="*", check.margin=FALSE))/2), nrow=P, ncol=Q)
     }
 
   # Column Shrinkage
-    .sim_delta1  <- function(Q, P, alpha.d1, delta.1, beta.d1, tau, sum.term, sigma = 1L) {
-        stats::rgamma(1, shape=alpha.d1 + P * Q/2, rate=beta.d1 + (sigma * 0.5)/delta.1 * tau %*% sum.term)
+    .sim_delta1  <- function(Q, P.5, alpha.d1, delta.1, beta.d1, tau, sum.term, sigma = 1L) {
+        stats::rgamma(1, shape=alpha.d1 + P.5 * Q, rate=beta.d1 + (sigma * 0.5)/delta.1 * tau %*% sum.term)
     }
 
-    .sim_deltak  <- function(Q, P, k, alpha.d2, beta.d2, delta.k, tau.kq, sum.term.kq, sigma = 1L) {
-        stats::rgamma(1, shape=alpha.d2 + P/2 * (Q - k + 1L), rate=beta.d2 + (sigma * 0.5)/delta.k * tau.kq %*% sum.term.kq)
+    .sim_deltak  <- function(Q, P.5, k, alpha.d2, beta.d2, delta.k, tau.kq, sum.term.kq, sigma = 1L) {
+        stats::rgamma(1, shape=alpha.d2 + P.5 * (Q - k + 1L), rate=beta.d2 + (sigma * 0.5)/delta.k * tau.kq %*% sum.term.kq)
     }
 
   # Cluster Shrinkage
-    .sim_sigma   <- function(G, P, Qs, rho1, rho2, sum.terms, tau) {
-        stats::rgamma(G, shape=rho1 + (P * Qs)/2, rate=rho2 + mapply("%*%", sum.terms, tau)/2)
+    .sim_sigma   <- function(G, P.5, Qs, rho1, rho2, sum.terms, tau) {
+        stats::rgamma(G, shape=rho1 + P.5 * Qs, rate=rho2 + mapply("%*%", sum.terms, tau)/2)
     }
 
   # Mixing Proportions
@@ -575,8 +575,8 @@
         exp.seq  <- lapply(ML, function(i) exp.Q1[i] * exp.Qk[i]^Qseq)
         check    <- !vapply(exp.seq, is.unsorted, logical(1L))
       }
-      exp.seq    <- if(length(exp.seq) == 1) exp.seq[[1L]]             else exp.seq
-      res        <- list(expectation = exp.seq, valid = if(Q < 2) TRUE else check)
+      exp.seq    <- if(length(exp.seq) == 1) exp.seq[[1L]] else exp.seq
+      res        <- list(expectation = exp.seq, valid = Q < 2 || check)
       attr(res, "Warning")    <- WX
         return(res)
     }
@@ -1044,23 +1044,25 @@
 #' @param N The sample size.
 #' @param alpha The concentration parameter. Must be specified (though not for \code{G_calibrate}) and must be strictly greater than \code{-discount}. The case \code{alpha=0} is accommodated. When \code{discount} is negative \code{alpha} must be a positive integer multiple of \code{abs(discount)}. See \strong{Details} for behaviour for \code{G_calibrate}.
 #' @param discount The discount parameter for the Pitman-Yor process. Must be less than 1, but typically lies in the interval [0, 1). Defaults to 0 (i.e. the Dirichlet process). When \code{discount} is negative \code{alpha} must be a positive integer multiple of \code{abs(discount)}. See \strong{Details} for behaviour for \code{G_calibrate}.
-#' @param MPFR Logical indicating whether the high-precision libraries \code{\link[Rmpfr]{Rmpfr}} and \code{gmp} are invoked, at the expense of run-time. Defaults to \code{TRUE} and \strong{must} be \code{TRUE} for \code{\link{G_expected}} when \code{alpha=0} and \code{\link{G_variance}} when \code{discount} is non-zero. See \strong{\code{Note}}.
-#' @param EG The prior expected number of clusters. Must exceed \code{1}.
+#' @param MPFR Logical indicating whether the high-precision libraries \code{\link[Rmpfr]{Rmpfr}} and \code{gmp} are invoked, at the expense of run-time. Defaults to \code{TRUE} and \strong{must} be \code{TRUE} for \code{G_expected} when \code{alpha=0} or \code{G_variance} when \code{discount} is non-zero. For \code{G_calibrate}, it is \emph{strongly recommended} to use \code{MPFR=TRUE} when \code{discount} is non-zero and strictly necessary when \code{alpha=0} is supplied. See \strong{\code{Note}}.
+#' @param EG The prior expected number of clusters. Must exceed \code{1} and be less than \code{N}.
 #' @param ... Additional arguments passed to \code{\link[stats]{uniroot}}, e.g. \code{maxiter}.
 #'
 #' @details All arguments are vectorised. Users can also consult \code{\link{G_priorDensity}} in order to solicit sensible priors.
 #'
-#' For \code{G_calibrate}, \strong{only one} of \code{alpha} or \code{discount} can be supplied, and the function elicits a value for the opposing parameter which achieves the desired expected number of clusters \code{EG} for the given sample size \code{N}. By default, a value for \code{alpha} subject to \code{discount=0} (i.e. the Dirichlet process) is elicited. See \strong{Examples} below.
+#' For \code{G_calibrate}, \strong{only one} of \code{alpha} or \code{discount} can be supplied, and the function elicits a value for the opposing parameter which achieves the desired expected number of clusters \code{EG} for the given sample size \code{N}. By default, a value for \code{alpha} subject to \code{discount=0} (i.e. the Dirichlet process) is elicited. Note that \code{alpha} may not be a positive integer multiple of \code{discount} as it should be if \code{discount} is negative. See \strong{Examples} below.
 #' @return The expected number of clusters under the specified prior conditions (\code{G_expected}), or the variance of the number of clusters (\code{G_variance}), or the concentration parameter \code{alpha} \strong{or} \code{discount} parameter achieving a particular expected number of clusters (\code{G_calibrate}).
 #' @keywords utility
 #' @export
 #' @name G_moments
 #' @rdname G_moments
 #'
-#' @note \code{G_variance} requires use of the \code{\link[Rmpfr]{Rmpfr}} and \code{gmp} libraries for non-zero \code{discount} values. \code{G_expected} requires these libraries only for the \code{alpha=0} case. Despite the high precision arithmetic used, the functions can still be unstable for small values of \code{discount}. See the argument \code{MPFR}.
+#' @note \code{G_variance} requires use of the \code{\link[Rmpfr]{Rmpfr}} and \code{gmp} libraries for non-zero \code{discount} values. \code{G_expected} requires these libraries only for the \code{alpha=0} case. These libraries are \emph{strongly recommended} (but they are not required) for \code{G_calbirate} when \code{discount} is non-zero, but they are required when \code{alpha=0} is supplied. Despite the high precision arithmetic used, the functions can still be unstable for large \code{N} and/or extreme values of \code{alpha} and/or \code{discount}. See the argument \code{MPFR}.
 #'
 #' @seealso \code{\link{G_priorDensity}}, \code{\link[Rmpfr]{Rmpfr}}, \code{\link[stats]{uniroot}}
 #' @references De Blasi, P., Favaro, S., Lijoi, A., Mena, R. H., Prunster, I., and Ruggiero, M. (2015) Are Gibbs-type priors the most natural generalization of the Dirichlet process?, \emph{IEEE Transactions on Pattern Analysis and Machine Intelligence}, 37(2): 212-229.
+#'
+#' Yamato, H. and Shibuya, M. (2000) Moments of some statistics of Pitman sampling formula, \emph{Bulletin of Informatics and Cybernetics}, 32(1): 1-10.
 #'
 #' @author Keefe Murphy - <\email{keefe.murphy@@mu.ie}>
 #' @usage
@@ -1087,54 +1089,48 @@
 #' # matplot(PY, type="l", xlab="N", ylab="G")
 #'
 #' # Other special cases of the PYP are also facilitated
-#' # G_expected(N=50, alpha=c(27.1401, 0), discount=c(-27.1401/100, 0.8054447))
-#' # G_variance(N=50, alpha=c(27.1401, 0), discount=c(-27.1401/100, 0.8054447))
+#' # G_expected(N=50, alpha=c(27.1401, 0), discount=c(-27.1401/100, 0.8054448))
+#' # G_variance(N=50, alpha=c(27.1401, 0), discount=c(-27.1401/100, 0.8054448))
 #'
-#' # Elicit values for alpha
+#' # Elicit values for alpha under a DP prior
 #' G_calibrate(N=50, EG=25)
-#' G_calibrate(N=50, EG=25, discount=c(0.25, 0.7300045))
 #'
-#' # Elicit values for discount
-#' G_calibrate(N=50, EG=25, alpha=c(12.21619, 1))
+#' # Elicit values for alpha under a PYP prior
+#' # require("Rmpfr")
+#' # G_calibrate(N=50, EG=25, discount=c(-27.1401/100, 0.25, 0.7300045))
+#'
+#' # Elicit values for discount under a PYP prior
+#' # G_calibrate(N=50, EG=25, alpha=c(12.21619, 1, 0), maxiter=2000)
     G_expected   <- Vectorize(function(N, alpha, discount = 0, MPFR = TRUE) {
       if(!all(is.numeric(N), is.numeric(discount),
          is.numeric(alpha)))               stop("All inputs must be numeric",     call.=FALSE)
       if(discount >= 1)                    stop("'discount' must be less than 1", call.=FALSE)
       if(discount > 0    &&
-         alpha   <= - discount)            stop("'alpha' must be strictly greater than -discount",    call.=FALSE)
+         alpha   <= - discount)            stop("'alpha' must be strictly greater than -discount",     call.=FALSE)
       if(discount < 0    &&
         (alpha   <= 0    ||
         !.IntMult(alpha,    discount)))    stop("'alpha' must be a positive integer multiple of 'abs(discount)' when 'discount' is negative", call.=FALSE)
-      if(alpha   == 0    && discount <= 0) stop("'discount' must be strictly positive when 'alpha=0", call.=FALSE)
+      if(alpha   == 0    && discount <= 0) stop("'discount' must be strictly positive when 'alpha'=0", call.=FALSE)
       if(alpha   == 0    && !isTRUE(MPFR)) stop("'MPFR' must be TRUE when 'alpha' == 0", call.=FALSE)
       igmp       <- isNamespaceLoaded("Rmpfr")
       if(mpfrind <- (isTRUE(MPFR)    &&
-                     suppressMessages(requireNamespace("Rmpfr", quietly=TRUE))    &&
-                     .version_above("gmp", "0.5-4"))) {
+                     suppressMessages(requireNamespace("Rmpfr", quietly=TRUE)) &&
+                     .version_above("gmp", "0.5-4")))       {
         if(isFALSE(igmp)) {
           on.exit(.detach_pkg("Rmpfr"))
           on.exit(.detach_pkg("gmp"), add=TRUE)
         }
         alpha    <- Rmpfr::mpfr(alpha, precBits=256)
-      }
+      }   else if(isTRUE(MPFR))            warning("'Rmpfr' package not installed\n", call.=FALSE, immediate.=TRUE)
 
       if(alpha    == 0)   {
         if(mpfrind)       {
-          tmp    <- sum(log(alpha + seq_len(N - 1L)))
-          ldisc  <- log(discount)
-          res    <- 0
-          for(k in seq_len(N)) {
-            kseq <- seq_len(k)
-            res  <- res   + k  * exp(sum(log(alpha + discount * kseq[-k])) - tmp
-                          - k  * ldisc) * sum((-1L)^kseq * gmp::chooseZ(n=k, k=kseq)
-                          * Rmpfr::pochMpfr(-kseq  * discount, N) / gmp::factorialZ(k))
-          }
-            return(gmp::asNumeric(res))
-        } else                           stop("'Rmpfr' must be installed when 'alpha'=0", call.=FALSE)
+          return(gmp::asNumeric(Rmpfr::pochMpfr(discount + 1, N - 1L)/gamma(Rmpfr::mpfr(N, precBits=256))))
+        } else                             stop("'Rmpfr' must be installed when 'alpha'=0", call.=FALSE)
       }
       if(discount == 0)   {
         exp      <- alpha * (digamma(alpha + N) - digamma(alpha))
-       #exp      <- sum(alpha/(alpha + 0L:(N - 1L)))
+       #exp      <- sum(alpha/(alpha  +  0L:(N  - 1L)))
         if(mpfrind)       {
           gmp::asNumeric(exp)
         } else    {
@@ -1145,7 +1141,7 @@
         if(mpfrind)       {
           gmp::asNumeric(adx  * Rmpfr::pochMpfr(alpha + discount, N)/Rmpfr::pochMpfr(alpha, N) - adx)
         } else    {
-          adx * (prod(discount/(alpha  + 0L:(N - 1L)) + 1L)   - 1L)
+          adx * (prod(discount/(alpha + 0L:(N - 1L))  + 1L) - 1L)
         }
       }
     })
@@ -1163,34 +1159,40 @@
          is.numeric(alpha)))               stop("All inputs must be numeric",     call.=FALSE)
       if(discount >= 1)                    stop("'discount' must be less than 1", call.=FALSE)
       if(discount > 0    &&
-         alpha   <= - discount)            stop("'alpha' must be strictly greater than -discount", call.=FALSE)
+         alpha   <= - discount)            stop("'alpha' must be strictly greater than -discount",     call.=FALSE)
       if(discount < 0    &&
         (alpha   <= 0    ||
         !.IntMult(alpha,    discount)))    stop("'alpha' must be a positive integer multiple of 'abs(discount)' when 'discount' is negative", call.=FALSE)
-     #if(alpha   == 0)    {                warning("'alpha'=0 case not yet implemented", call.=FALSE, immediate.=TRUE); return(Inf) }
-      if(discount != 0   && !isTRUE(MPFR)) stop("'MPFR' must be TRUE when 'discount' is non-zero", call.=FALSE)
+      if(alpha   == 0    && discount <= 0) stop("'discount' must be strictly positive when 'alpha'=0", call.=FALSE)
+      if(discount != 0   && !isTRUE(MPFR)) stop("'MPFR' must be TRUE when 'discount' is non-zero",     call.=FALSE)
       igmp       <- isNamespaceLoaded("Rmpfr")
       if(mpfrind <- (isTRUE(MPFR)    &&
-                     suppressMessages(requireNamespace("Rmpfr", quietly=TRUE))    &&
+                     suppressMessages(requireNamespace("Rmpfr", quietly=TRUE)) &&
                      .version_above("gmp", "0.5-4"))) {
         if(isFALSE(igmp)) {
           on.exit(.detach_pkg("Rmpfr"))
           on.exit(.detach_pkg("gmp"), add=TRUE)
         }
-        alpha    <- Rmpfr::mpfr(alpha, precBits=256)
-      } else if(discount != 0)             stop("'Rmpfr' package not installed", call.=FALSE)
+        alpha    <- Rmpfr::mpfr(alpha,   precBits=256)
+      }   else    {
+        if(discount      != 0)        {    stop("'Rmpfr' package not installed",      call.=FALSE)
+        } else if(isTRUE(MPFR))            warning("'Rmpfr' package not installed\n", call.=FALSE, immediate.=TRUE)
+      }
 
       alpha2     <- alpha * alpha
       if(discount == 0)   {
         var      <- alpha * (digamma(alpha  + N) - digamma(alpha))
         if(mpfrind)       {
           alpha  <- gmp::asNumeric(alpha)
-          gmp::asNumeric(var  + alpha2 * (trigamma(alpha + N) - trigamma(alpha)))
+          gmp::asNumeric(var + alpha2 * (trigamma(alpha + N) - trigamma(alpha)))
         } else {
           var  + alpha2   * (trigamma(alpha + N) - trigamma(alpha))
         }
+      } else if(alpha    == 0) {
+        poch.1   <- gamma(Rmpfr::mpfr(N, precBits=256))
+        subterm  <- Rmpfr::pochMpfr(discount + 1, N - 1L)/poch.1
+          gmp::asNumeric(Rmpfr::pochMpfr(2 * discount, N)/(discount * poch.1) - subterm  * (1L + subterm))
       } else   {
-        alpha    <- if(alpha == 0) .Machine$double.eps else alpha
         sum.ad   <- alpha + discount
         poch.a   <- Rmpfr::pochMpfr(alpha, N)
         poch.ad  <- Rmpfr::pochMpfr(sum.ad, N)
@@ -1204,38 +1206,72 @@
 #' @usage
 #' G_calibrate(N,
 #'             EG,
-#'             discount = 0,
 #'             alpha = NULL,
+#'             discount = 0,
+#'             MPFR = TRUE,
 #'             ...)
 #' @export
-    G_calibrate  <- Vectorize(function(N, EG, discount = 0, alpha = NULL, ...) {
+    G_calibrate  <- Vectorize(function(N, EG, alpha = NULL, discount = 0, MPFR = TRUE, ...) {
       if(!all(is.numeric(N), is.numeric(discount), is.null(alpha) || is.numeric(alpha),
          is.numeric(EG)))                  stop("All inputs must be numeric",     call.=FALSE)
       if(discount >= 1)                    stop("'discount' must be less than 1", call.=FALSE)
-      if(EG       <= 1)                    stop("'EG' must be greater than 1",   call.=FALSE)
-      RFA        <- function(N, alpha, discount) prod(1 + discount/(alpha + 0L:(N - 1L)))
-      if(is.null(alpha))    {
-        if(discount == 0)   {
-          X      <- try(suppressWarnings(stats::uniroot(function(x) sum(x/(x   + 0L:(N - 1L)))             - EG, interval=c(0.00001, 10000), ...)),             silent=TRUE)
-        } else    {
-          X      <- try(suppressWarnings(stats::uniroot(function(x) x/discount * (RFA(N, x, discount) - 1) - EG, interval=c(-discount + 0.00001, 10000), ...)), silent=TRUE)
+      if(EG       <= 1)                    stop("'EG' must be greater than 1",    call.=FALSE)
+      if(EG       >= N)                    stop("'EG' must be less than 'N'",     call.=FALSE)
+      igmp       <- isNamespaceLoaded("Rmpfr")
+      if(mpfrind <- (isTRUE(MPFR)    &&
+                     suppressMessages(requireNamespace("Rmpfr", quietly=TRUE)) &&
+                     .version_above("gmp", "0.5-4"))) {
+        if(isFALSE(igmp))   {
+          on.exit(.detach_pkg("Rmpfr"))
+          on.exit(.detach_pkg("gmp"), add=TRUE)
         }
-        if(inherits(X, "try-error")) {     warning("uniroot failed to elicit an alpha value\n",   call.=FALSE, immediate.=TRUE)
-          Y      <- stats::setNames(NA,      "alpha")
-        } else Y <- stats::setNames(X$root,  "alpha")
+      } else if(isTRUE(MPFR))              warning("'Rmpfr' package not installed\n", call.=FALSE, immediate.=TRUE)
+      if(isTRUE(mpfrind)   &&
+        (!is.null(alpha)   || discount != 0))          {
+        if(!is.null(alpha) && alpha    == 0)           {
+          p1     <- gamma(Rmpfr::mpfr(N, precBits=256))
+          RFA    <- function(N, discount, p1)          {
+            gmp::asNumeric(Rmpfr::pochMpfr(discount + 1, N - 1L)/p1)
+          }
+        } else    {
+          RFA    <- function(N, alpha, discount)       {
+            x    <- gmp::asNumeric((Rmpfr::pochMpfr(alpha  + discount, N)/Rmpfr::pochMpfr(alpha, N)  - 1)    * alpha/discount)
+            ifelse(x == 0, N, ifelse(is.infinite(x)   && x < 0, 1, x))
+          }
+        }
+      } else RFA <- function(N, alpha, discount) alpha/discount * (prod(1 + discount/(alpha  + 0L:(N - 1L))) - 1)
+
+      if(is.null(alpha))    {
+        if(discount  == 0)  {
+          inter  <- c(.Machine$double.eps,    .Machine$double.xmax)
+          X      <- try(suppressWarnings(stats::uniroot(function(x) sum(x/(x + 0L:(N - 1L))) - EG, interval=inter, ...)), silent=TRUE)
+        } else    {
+          inter  <- if(isTRUE(mpfrind)) c(-discount + .Machine$double.eps, .Machine$double.xmax) else c(-discount + 0.000001,  100000)
+          X      <- try(suppressWarnings(stats::uniroot(function(x) RFA(N, x, discount)      - EG, interval=inter, ...)), silent=TRUE)
+        }
+        if(inherits(X, "try-error")) {     warning(paste0("uniroot failed to elicit a discount value", ifelse(isFALSE(MPFR), ": consider setting MPFR=TRUE\n","\n")), call.=FALSE, immediate.=TRUE)
+          Y      <- stats::setNames(NA,         "alpha")
+        } else Y <- stats::setNames(X$root,     "alpha")
       }   else if(missing(discount) ||
                   discount == 0)     {
-        if(alpha == 0)                     warning("'alpha'=0 case not yet implemented\n",        call.=FALSE, immediate.=TRUE)
-        X        <- try(suppressWarnings(stats::uniroot(function(x) alpha/x    * (RFA(N, alpha, x)    - 1) - EG, interval=c(-10000, 1 - 0.00001), ...)),        silent=TRUE)
-        if(inherits(X, "try-error")) {     warning("uniroot failed to elicit a discount value\n", call.=FALSE, immediate.=TRUE)
+        if(alpha == 0)               {
+          if(!isTRUE(MPFR))                stop("'MPFR' must be TRUE when 'alpha' == 0",          call.=FALSE)
+          inter  <- c(.Machine$double.eps,   1 - .Machine$double.eps)
+          X      <- try(suppressWarnings(stats::uniroot(function(x) RFA(N, x,    p1)         - EG, interval=inter, ...)), silent=TRUE)
+        } else    {
+          inter  <- c(-.Machine$double.xmax, 1 - .Machine$double.eps)
+          X      <- try(suppressWarnings(stats::uniroot(function(x) RFA(N, alpha, x)         - EG, interval=inter, ...)), silent=TRUE)
+        }
+        if(inherits(X, "try-error")) {     warning(paste0("uniroot failed to elicit a discount value", ifelse(isFALSE(MPFR), ": consider setting MPFR=TRUE\n","\n")), call.=FALSE, immediate.=TRUE)
           Y      <- stats::setNames(NA,      "discount")
         } else Y <- stats::setNames(X$root,  "discount")
       }   else                             stop("'alpha' and 'discount' cannot both be supplied", call.=FALSE)
       dots       <- list(...)
       maxiter    <- ifelse(length(dots) > 0 && any(names(dots) %in% "maxiter"), dots$maxiter, 1000)
-      if(X$iter  == maxiter)               warning(paste0("uniroot failed to converge in ", maxiter, " iterations\n"), call.=FALSE)
+      if(!inherits(X, "try-error")  &&
+         X$iter  == maxiter)               warning(paste0("uniroot failed to converge in ", maxiter, " iterations\n"), call.=FALSE)
         return(Y)
-    },  vectorize.args = c("N", "EG", "discount", "alpha"))
+    },  vectorize.args = c("N", "EG", "alpha", "discount", "MPFR"))
 
   # Print functions
 #' @method print IMIFA
@@ -2392,8 +2428,7 @@
 
     .version_above        <- function(pkg, than) {
       pkg        <- as.character(utils::packageVersion(pkg))
-      test       <- ifelse(test <- identical(pkg, than), test, as.logical(utils::compareVersion(pkg, than)))
-        return(test)
+        identical(pkg, than) || (utils::compareVersion(pkg, than) >= 0)
     }
 
     .which0      <- function(x) which(x == 0)